requirements.txt
Filetype==1.0.7
Pillow==8.0.1
PyMuPDF==1.18.9 
pytesseract==0.3.7